Using typography in chip component + type cleanup
