+++ lib/components/AriaAnchor.tsx
@@ -10,7 +10,7 @@ type AnchorType = {
   href: string
 }
 
-const AriaAnchor: React.FC<AnchorType> = forwardRef<HTMLAnchorElement, AnchorType>(
+const AriaAnchor = forwardRef<HTMLAnchorElement, AnchorType>(
   ({ children, className, forwardedRef, ...restProps }, ref) => {
     const { linkProps } = useLink({ ...restProps }, ref as RefObject<HTMLAnchorElement>)
