+++ lib/types/tagGroupTypes.ts
@@ -1,17 +1,27 @@
-export type ChipSizeType = 'large' | 'medium' | 'small'
+import { Selection } from 'react-aria-components'
 
-export type ChipColorsType = 'blue' | 'brown' | 'green' | 'purple' | 'red' | 'yellow' | 'disabled'
+export type ChipSizeProps = 'large' | 'medium' | 'small'
 
-export type ChipTagType = {
+export type ChipColorsProps = 'blue' | 'brown' | 'green' | 'purple' | 'red' | 'yellow' | 'disabled'
+
+export type ChipProps = {
   key: string
   id: string
   label: string | React.ReactNode
-  size?: ChipSizeType
+  size?: ChipSizeProps
   chipStyle?: string
-  color?: ChipColorsType
+  color?: ChipColorsProps
 }
 
-export type ChipProps = {
-  tags: ChipTagType[]
+export type BaChipProps = {
+  tags: ChipProps[]
   children?: React.ReactNode
+}
+
+export type BaTagGroupProps = {
+  setTags: (keys: Selection) => void
+  selectedTags: Selection
+  disabledTags?: string[]
+  defaultSelectedTags?: Iterable<string>
+  tagListClassName?: string
 }
